Fix makefiles and be MSVC++ compliant for Utils

+++ Utilities/include/Network/Socket.h
@@ -74,7 +74,7 @@ namespace Stardust::Network {
 		/**
 		 * Threshold for zlib compression.
 		**/
-		int threshold;
+		size_t threshold;
 	};
 
 	/**
